+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Ongoing
 	- These three classes can now be mixed.
 
 - Added :class:`~.SanStream` for non-waste streams (e.g., gases).
+- Updated the ``add_OPEX`` attribute of :class:`~.SanUnit` and ``system_add_OPEX`` attribute of :class:`~.SimpleTEA` so that they take :class:`dict` as the default to allow display of multiple additional operating expenses.
 - Split the ``systems`` module into an individual package `EXPOsan <https://github.com/QSD-Group/exposan>`_.